% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{package}
\name{MergeBy}
\alias{MergeBy}
\alias{MergeBy-package}
\title{MergeBy.}
\usage{
MergeBy(x, y)
}
\arguments{
\item{x}{An R data frame or object coercible into a data frame}

\item{y}{An R data frame or object coercible into a data frame which you wish
to merge with x}
}
\description{
MergeBy.

This function helps find common columns within (potentially large) dataframes
}
\examples{
MergeBy(mydf0, mydf1)
MergeBy()
}
\keyword{bind}
\keyword{join}
\keyword{merge}

